Moved to Maven Structure
