+++ .idea/compiler.xml
@@ -6,11 +6,11 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="TeamSpeak" />
+        <module name="TeamSpeakIP" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
-      <module name="TeamSpeak" target="1.8" />
+      <module name="TeamSpeakIP" target="1.8" />
     </bytecodeTargetLevel>
   </component>
 </project>
