+++ .idea/encodings.xml
@@ -1,6 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/src" charset="ISO-8859-1" />
+  <component name="Encoding" defaultCharsetForPropertiesFiles="UTF-8">
+    <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/de/t0biii/ts/TeamSpeak.java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/de/t0biii/ts/listener/PlayerJoin.java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/de/t0biii/ts/methods/HelpMessages.java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/de/t0biii/ts/methods/files/Messages.java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
+    <file url="PROJECT" charset="UTF-8" />
   </component>
 </project>
