+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/TeamSpeak.iml" filepath="$PROJECT_DIR$/TeamSpeak.iml" />
+      <module fileurl="file://$PROJECT_DIR$/TeamSpeakIP.iml" filepath="$PROJECT_DIR$/TeamSpeakIP.iml" />
     </modules>
   </component>
 </project>
