--- plugin.yml
@@ -1,12 +1,0 @@
-name: TeamSpeakIP
-author: T0biii
-version: 5.7.3
-website: http://dev.bukkit.org/bukkit-plugins/teamspeak-ip/
-description: Shows you the Teamspeak3IP
-prefix: TeamSpeakIP 
-
-main: de.t0biii.ts.TeamSpeak
-commands:
-  ts:
-    description: Shows the TS IP
-    aliases: [teamspeak, ts3, teamspeakip]
