+++ src/main/java/de/t0biii/ts/methods/JsonMessage.java
