+++ src/main/resources/plugin.yml
@@ -1,0 +1,13 @@
+name: TeamSpeakIP
+author: T0biii
+version: ${project.version}
+website: http://dev.bukkit.org/bukkit-plugins/teamspeak-ip/
+description: Shows you the Teamspeak3IP
+prefix: TeamSpeakIP 
+
+main: de.t0biii.ts.TeamSpeak
+api-version: 1.13
+commands:
+  ts:
+    description: Shows the TS IP
+    aliases: [teamspeak, ts3, teamspeakip]
